+++ packages/core/src/components/autosuggest/autosuggest.tsx
@@ -15,9 +15,22 @@ import debounce from "debounce";
 import { v4 } from "uuid";
 import escapeStringRegexp from "escape-string-regexp";
 
+// Keep in sync with readme.md
 export interface Suggestion {
+  /**
+   * The text that will be displayed as the suggestion.
+   */
   value: string;
+
+  /**
+   * The type of suggestion.
+   */
   type?: string;
+
+  /**
+   * A reference to the original object that was used to create the suggestion.
+   */
+  item?: unknown;
 }
 
 @Component({
@@ -27,7 +40,11 @@ export interface Suggestion {
 })
 export class Autosuggest {
   /**
-   * The suggestions for the value of the slotted input element
+   * The suggestions for the value of the slotted input element. Optionally a
+   * Suggestion can have a `type` and `item`.
+   * 
+   * The `type` is used to style the suggestion. `item` can be use to reference
+   * the original object that was used to create the suggestion.
    */
   @Prop()
   readonly suggestions: Suggestion[] = [];