+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 * **css:** Nieuw component "Tile Grid" en "Tile" uitgebreid met variant "theme" ([#1443](https://github.com/dso-toolkit/dso-toolkit/issues/1443))
 * **core:** Nieuw component "Image Overlay": `<dso-image-overlay>` ([#1444](https://github.com/dso-toolkit/dso-toolkit/issues/1444))
+* **core:** Autosuggest uitbreiding voor verwijzing naar object bij suggestie ([#1494](https://github.com/dso-toolkit/dso-toolkit/issues/1494))
 
 ### Changed
 * **css + dso-toolkit + sources:** "Rich Content" verhuizen naar `@dso-toolkit/sources` ([#1465](https://github.com/dso-toolkit/dso-toolkit/issues/1465))